Rename files, move KeyedArray class
